notification-service/requirements.txt
# Notification Service Dependencies

fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Kafka
confluent-kafka==2.3.0

# Monitoring
prometheus-fastapi-instrumentator

# Distributed Tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-jaeger-thrift==1.21.0

# For production notification providers (optional)
# sendgrid==6.10.0
# twilio==8.10.0
# firebase-admin==6.3.0